WIP: Move `Menu*` interface defs from `constants.ts` to new `types.ts` file
