+++ packages/mdc-menu-surface/foundation.ts
@@ -23,7 +23,8 @@
 
 import MDCFoundation from '@material/base/foundation';
 import {MDCMenuSurfaceAdapter} from './adapter';
-import {Corner, CornerBit, cssClasses, MenuDimensions, MenuDistance, MenuPoint, numbers, strings} from './constants';
+import {Corner, CornerBit, cssClasses, numbers, strings} from './constants';
+import {MenuDimensions, MenuDistance, MenuPoint} from './types';
 
 interface AutoLayoutMeasurements {
   anchorSize: MenuDimensions;
